Updated to use events for the service.
