+++ .vscode/launch.json
@@ -4,17 +4,25 @@
     // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
     "version": "0.2.0",
     "configurations": [
+        // {
+        //     "type": "node",
+        //     "request": "launch",
+        //     "name": "Launch Program",
+        //     "skipFiles": [
+        //         "<node_internals>/**"
+        //     ],
+        //     "program": "${workspaceFolder}\\src\\repl.ts",
+        //     "outFiles": [
+        //         "${workspaceFolder}/**/*.js"
+        //     ],
+        //     "console": "integratedTerminal"
+        // }
         {
             "type": "node",
             "request": "launch",
-            "name": "Launch Program",
-            "skipFiles": [
-                "<node_internals>/**"
-            ],
-            "program": "${workspaceFolder}\\src\\index.ts",
-            "outFiles": [
-                "${workspaceFolder}/**/*.js"
-            ],
+            "name": "Launch via typescript",
+            "program": "${workspaceFolder}\\src\\repl.ts",
+            "outFiles": ["${workspaceFolder}/**/*.js"],
             "console": "integratedTerminal"
         }
     ]
