+++ src/domain/ScriptState.ts
@@ -1,0 +1,8 @@
+// Exports
+export interface ScriptState
+{
+    name : string,
+    guid : string,
+    script : string,
+    ui : string
+}
