+++ tsconfig.json
@@ -2,11 +2,12 @@
   "include": ["src", "types", "test"],
   "compilerOptions": {
     "target": "es5",
-    "module": "esnext",
+    "module": "commonjs",
     "lib": ["dom", "esnext"],
     "importHelpers": true,
     "declaration": true,
     "sourceMap": true,
+    "outDir": "./dist",
     "rootDir": "./",
     "strict": true,
     "noImplicitAny": true,
